Update index.js
added comments, replace throw process.exit(1) with methods processFault and processLoopOnFault with log the fault message every 5 seconds. This was a better option since exiting would cause the accessory to be constantly be reloaded and this approach allowed the error message to be easily seen in the log and prevents automatic reloading.